requirements.txt
dnspython>=2.4.2
colorama>=0.4.6
pyfiglet>=1.0.2
rich>=13.7.0
requests>=2.31.0
cryptography>=42.0.2
pyOpenSSL>=24.0.0
python-dateutil>=2.8.2 